+++ lib/src/client/storage_impl.dart
@@ -61,8 +61,14 @@ class LocalstorageCargo extends Cargo {
     return values.length;
   }
 
-  Map export() {
+  Map exportSync() {
     return values;
+  }
+  
+  Future<Map> export() {
+    Completer complete = new Completer();
+    complete.complete(values);
+    return complete.future;
   }
 
   Future start() => _completer.future;